fix(columbia_merger): move functions to corpus_utils
many of these functions can be reused in columbia importer
